refactor: 커서 생성에 필요한 DTO 패키지 경로 이동
